+++ src/index.test.ts
@@ -146,7 +146,7 @@ describe("react-declassify", () => {
       `;
       const output = dedent`\
         import React from "react";
-        const C = () => {};
+        const C = React.memo(function C() {});
       `;
       expect(transform(input)).toBe(output);
     });
@@ -1018,6 +1018,40 @@ describe("react-declassify", () => {
         };
       `;
       expect(transform(input)).toBe(output);
+    });
+  });
+
+  describe("Memoization", () => {
+    it("transforms PureComponent to React.memo", () => {
+      const input = dedent`\
+        class C extends React.PureComponent {
+          render() {
+            return <div>Hello, world!</div>;
+          }
+        }
+      `;
+      const output = dedent`\
+        const C = React.memo(function C() {
+          return <div>Hello, world!</div>;
+        });
+      `;
+      expect(transform(input)).toBe(output);
+    });
+
+    it("Places types on const", () => {
+      const input = dedent`\
+        class C extends React.PureComponent {
+          render() {
+            return <div>Hello, world!</div>;
+          }
+        }
+      `;
+      const output = dedent`\
+        const C: React.FC = React.memo(function C() {
+          return <div>Hello, world!</div>;
+        });
+      `;
+      expect(transform(input, { ts: true })).toBe(output);
     });
   });